Connect signals through editor instead of code
Connecting signals from timers via the editor instead of code has the same effect,
but instead of connecting the signals at run time to the mole callbacks
functions, it keeps this information in the `.tscn` file and removes the
code which would otherwise be adding clutter. Connecting signals via code is
useful in the event we want to create an object on the fly and there is
no dedicated node in the scene tree.